refactor(typescript): extract HTTP Range helpers into internal/httpRange
Move fetchRange and the FetchRangeContext shape out of file.ts into a
new internal module so the forthcoming layout-aware backend
(messageMetadata / messageObject / prefetchLayouts) can reuse the same
206 / 200 / eager-fallback behaviour without duplicating it.

No behaviour change — existing tests cover the same code paths, just
reached through an internal import.  File shrinks by ~60 lines;
internal API surface grows by one exported function and one interface
(both `@internal`, not re-exported from index.ts).
